+++ rds-20140815/alibabacloud_rds20140815/models/_modify_rcdisk_attribute_request.py
@@ -4,25 +4,23 @@ from __future__ import annotations
 
 from darabonba.model import DaraModel
 
-class DetachRCDiskRequest(DaraModel):
+class ModifyRCDiskAttributeRequest(DaraModel):
     def __init__(
         self,
+        bursting_enabled: bool = None,
         delete_with_instance: bool = None,
+        description: str = None,
         disk_id: str = None,
-        instance_id: str = None,
+        disk_name: str = None,
         region_id: str = None,
     ):
-        # The reserved parameter. This parameter is not supported.
+        self.bursting_enabled = bursting_enabled
         self.delete_with_instance = delete_with_instance
-        # The ID of the disk that you want to detach.
-        # 
+        self.description = description
         # This parameter is required.
         self.disk_id = disk_id
-        # The instance ID.
-        # 
+        self.disk_name = disk_name
         # This parameter is required.
-        self.instance_id = instance_id
-        # The region ID. You can call the DescribeRegions operation to query the most recent region list.
         self.region_id = region_id
 
     def validate(self):
@@ -33,14 +31,20 @@ class DetachRCDiskRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.bursting_enabled is not None:
+            result['BurstingEnabled'] = self.bursting_enabled
+
         if self.delete_with_instance is not None:
             result['DeleteWithInstance'] = self.delete_with_instance
+
+        if self.description is not None:
+            result['Description'] = self.description
 
         if self.disk_id is not None:
             result['DiskId'] = self.disk_id
 
-        if self.instance_id is not None:
-            result['InstanceId'] = self.instance_id
+        if self.disk_name is not None:
+            result['DiskName'] = self.disk_name
 
         if self.region_id is not None:
             result['RegionId'] = self.region_id
@@ -49,14 +53,20 @@ class DetachRCDiskRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('BurstingEnabled') is not None:
+            self.bursting_enabled = m.get('BurstingEnabled')
+
         if m.get('DeleteWithInstance') is not None:
             self.delete_with_instance = m.get('DeleteWithInstance')
+
+        if m.get('Description') is not None:
+            self.description = m.get('Description')
 
         if m.get('DiskId') is not None:
             self.disk_id = m.get('DiskId')
 
-        if m.get('InstanceId') is not None:
-            self.instance_id = m.get('InstanceId')
+        if m.get('DiskName') is not None:
+            self.disk_name = m.get('DiskName')
 
         if m.get('RegionId') is not None:
             self.region_id = m.get('RegionId')